+++ COFantasy.js
@@ -1434,14 +1434,14 @@ var COFantasy = COFantasy || function() {
     if (perso.token && perso.token.get('bar1_link') === '') {
       msg = perso.token.get('name') + ' ' + msg;
       if (secret) {
-        var character = getObj('character', perso.charId);
+        let character = getObj('character', perso.charId);
         if (character) {
-          var controlled = character.get('controlledby');
+          let controlled = character.get('controlledby');
           if (controlled.includes('all')) sendChat('', msg);
           else {
             controlled.split(',').forEach(function(c) {
               if (c !== '' && !playerIsGM(c)) {
-                var p = getObj('player', c);
+                let p = getObj('player', c);
                 if (p && p.get('online')) {
                   sendChat('', '/w "' + p.get('_displayname') + '" ' + msg);
                 }
@@ -1666,12 +1666,12 @@ var COFantasy = COFantasy || function() {
   }
 
   function getValeurOfEffet(perso, effet, def, predDef) {
-    var attrsVal = tokenAttribute(perso, effet + 'Valeur');
+    let attrsVal = tokenAttribute(perso, effet + 'Valeur');
     if (attrsVal.length === 0) {
       if (predDef) return predicateAsInt(perso, predDef, def);
       return def;
     }
-    var res = parseInt(attrsVal[0].get('current'));
+    let res = parseInt(attrsVal[0].get('current'));
     if (isNaN(res)) return def;
     return res;
   }
@@ -1884,6 +1884,47 @@ var COFantasy = COFantasy || function() {
           case 'armeeDesMortsTempeteDeManaIntense':
             personnage.token.set("aura2_radius", Math.floor(20 * Math.sqrt(1 + value)));
             break;
+          case 'masqueDuPredateur':
+            {
+              //L'initiative change
+              initPerso(personnage, evt, true);
+              let lie = personnageAmeLiee(personnage);
+              if (lie) {
+                setTokenAttr(lie, 'masqueDuPredateurAmeLiee', value, evt, {
+                  maxVal: options.maxVal
+                });
+                let valeur = getValeurOfEffet(personnage, 'masqueDuPredateur', modCarac(personnage, 'sagesse'));
+                valeur = Math.floor(valeur / 2);
+                if (valeur > 1)
+                  setTokenAttr(lie, 'masqueDuPredateurAmeLieeValeur', valeur, evt);
+                initPerso(lie, evt, true);
+              }
+            }
+            break;
+          case 'masqueMortuaire':
+            {
+              let lie = personnageAmeLiee(personnage);
+              if (lie) {
+                setTokenAttr(lie, 'masqueMortuaireAmeLiee', value, evt, {
+                  maxVal: options.maxVal
+                });
+              }
+            }
+            break;
+          case 'peauDEcorce':
+            {
+              let lie = personnageAmeLiee(personnage);
+              if (lie) {
+                setTokenAttr(lie, 'peauDEcorceAmeLiee', value, evt, {
+                  maxVal: options.maxVal
+                });
+                let valeur = getValeurOfEffet(personnage, 'peauDEcorce', 1, 'voieDesVegetaux');
+                valeur = Math.floor(valeur / 2);
+                if (valeur > 1)
+                  setTokenAttr(lie, 'peauDEcorceAmeLieeValeur', valeur, evt);
+              }
+            }
+            break;
         }
       }
       return attr;
@@ -1995,8 +2036,8 @@ var COFantasy = COFantasy || function() {
     if (link === '') regSuivis += "_" + token.get('name') + '$';
     else regSuivis += '$';
     regSuivis = new RegExp(regSuivis);
-    var soinsSuivis = soins;
-    var soinsImpossible = new Set(options.saufDMType);
+    let soinsSuivis = soins;
+    let soinsImpossible = new Set(options.saufDMType);
     attrs.forEach(function(a) {
       if (soinsSuivis === 0) return;
       var an = a.get('name');
@@ -2153,7 +2194,7 @@ var COFantasy = COFantasy || function() {
   }
 
   function persoOfToken(token) {
-    var charId = token.get('represents');
+    let charId = token.get('represents');
     if (charId === '') {
       return undefined;
     }
@@ -2854,6 +2895,71 @@ var COFantasy = COFantasy || function() {
     }
   }
 
+  function compagnonPresent(personnage, attribut) {
+    let attrs = findObjs({
+      _type: 'attribute',
+      _characterid: personnage.charId,
+      name: attribut
+    });
+    if (attrs.length > 0) {
+      let compagnon = attrs[0].get('current');
+      let compToken = findObjs({
+        _type: 'graphic',
+        _subtype: 'token',
+        _pageid: personnage.token.get('pageid'),
+        layer: 'objects',
+        name: compagnon
+      });
+      let res = false;
+      compToken.forEach(function(tok) {
+        if (res) return;
+        let compCharId = tok.get('represents');
+        if (compCharId === '') return;
+        compagnon = {
+          token: tok,
+          charId: compCharId
+        };
+        res = isActive(compagnon);
+      });
+      return res;
+    }
+    return false;
+  }
+
+  function personnageAmeLiee(perso, pageId, allToks) {
+    let lie = predicateAsBool(perso, 'ameLieeAvec');
+    if (!lie) return;
+    let tokenLie;
+    if (allToks) {
+      tokenLie = allToks.find(function(tok) {
+        return tok.get('name') == lie;
+      });
+      if (tokenLie) return persoOfToken(tokenLie);
+    } else {
+      pageId = pageId || perso.token.get('pageid');
+      tokenLie = findObjs({
+        _type: 'graphic',
+        _pageid: pageId,
+        _subtype: 'token',
+        layer: 'objects',
+        name: lie
+      });
+      if (tokenLie.length > 0) {
+        let p;
+        let t = tokenLie.find(function(tok) {
+          if (tok.id == perso.token.id) return false;
+          p = persoOfToken(tok);
+          if (p === undefined) return false;
+          if (getState(p, 'mort')) return false;
+          return true;
+        });
+        if (t) return p;
+      } else {
+        return;
+      }
+    }
+  }
+
   //options:
   //fromTemp si on est en train de supprimer un effet temporaire
   function setState(personnage, etat, value, evt, options) {
@@ -3289,27 +3395,30 @@ var COFantasy = COFantasy || function() {
               });
             //On cherche d'abord si un siphon des âmes est prioritaire
             let prioriteSiphon = [];
-            allToks.forEach(function(tok) {
-              if (tok.id == token.id) return;
+            allToks = allToks.filter(function(tok) {
+              if (tok.id == token.id) return false;
               let p = persoOfToken(tok);
-              if (p === undefined) return;
-              if (getState(p, 'mort')) return;
-              if (distanceCombat(token, tok, pageId) > 20) return;
+              if (p === undefined) return false;
+              if (getState(p, 'mort')) return false;
+              if (distanceCombat(token, tok, pageId) > 20) return false;
               if (predicateAsBool(p, 'siphonDesAmes')) {
                 prioriteSiphon.push({
                   perso: p,
                   priorite: predicateAsInt(p, 'siphonDesAmesPrioritaire', 0)
                 });
               }
+              return true;
             });
             if (prioriteSiphon.length > 0) {
+              let pvMax = parseInt(personnage.token.get('bar1_max'));
+              if (isNaN(pvMax) || pvMax < 1) pvMax = 1;
               if (estPJ(personnage)) {
                 let siphoneur = prioriteSiphon[0].perso;
                 let bonus = predicateAsInt(siphoneur, 'siphonDesAmes', 0);
                 let jetSiphon = "(1d6";
                 if (bonus > 0) jetSiphon += '+' + bonus;
                 jetSiphon += ")";
-                sendChat('COF', "/w GM " + personnage.token.get('name') + " est un personnage joueur, possible qu'il ne soit pas vraiment mort mais juste inconscient. Si il est vraiment mort, faire le siphon des âmes par " + siphoneur.token.get('name') + " à la main " + jetSiphon);
+                sendChat('COF', "/w GM " + personnage.token.get('name') + " est un personnage joueur, possible qu'il ne soit pas vraiment mort mais juste inconscient. S'il est vraiment mort, faire le siphon des âmes par " + siphoneur.token.get('name') + " à la main " + jetSiphon);
               } else {
                 prioriteSiphon.sort(function(a, b) {
                   return b.priorite - a.priorite;
@@ -3328,26 +3437,74 @@ var COFantasy = COFantasy || function() {
                     return;
                   }
                   let bonus = predicateAsInt(p, 'siphonDesAmes', 0);
+                  let nbDes = 1 + Math.floor(pvMax / 50);
                   let soin = rollDePlus(6, {
-                    bonus: bonus
+                    bonus,
+                    nbDes
                   });
                   let soinTotal = soin.val;
+                  //Le montant total des soins ne peut excéder les pv max du personnage qui vient de mourrir.
+                  let display = true;
+                  if (soinTotal > pvMax) {
+                    soinTotal = pvMax;
+                    display = false;
+                  }
+                  if (soinTotal < 1) soinTotal = 1;
+                  soin.val = soinTotal;
                   soin.val = Math.ceil(soin.val * fractionPriorite / 100);
+                  //Cherche si il y a un perso lié
+                  let lie = personnageAmeLiee(p, pageId, allToks);
                   soigneToken(p, soin.val, evt,
                     function(s) {
                       let siphMsg = "siphone l'âme de " + token.get('name') +
                         ". " + onGenre(p, 'Il', 'Elle') + " récupère ";
                       if (s == soinTotal) {
-                        siphMsg += soin.roll + " pv.";
+                        if (display) siphMsg += soin.roll + " pv.";
+                        else siphMsg += s + " pv (jet " + soin.roll + ").";
                         fraction = 0;
                       } else {
                         siphMsg += s + " pv (jet " + soin.roll + ").";
                         fraction -= Math.ceil(s * 100 / soinTotal);
                       }
+                      pvMax -= s;
                       whisperChar(p.charId, siphMsg);
+                      if (pvMax > 0 && lie && (s > 4 || s < soinTotal)) {
+                        let soin2 = soinTotal - s + Math.floor(s / 5);
+                        soigneToken(lie, soin2, evt, function(s) {
+                          let siphMsg = "récupère une partie de l'âme de " + token.get('name') +
+                            "siphonée par " + nomPerso(p) + ". " + onGenre(lie, 'Il', 'Elle') + " récupère " + s + " pv.";
+                          if (s == soin2) {
+                            fraction = 0;
+                          } else {
+                            fraction -= Math.ceil(s * 100 / soin2);
+                          }
+                          pvMax -= s;
+                          whisperChar(lie.charId, siphMsg);
+                        }, function() {
+                          whisperChar(lie.charId, "sent qu'" + onGenre(lie, 'il', 'elle') + " aurait pu bénéficier d'une âme");
+                        });
+                      }
                     },
                     function() {
-                      whisperChar(p.charId, "est déjà au maximum de point de vie. Il laisse échapper l'âme de " + token.get('name'));
+                      if (lie) {
+                        soigneToken(lie, soin.val, evt, function(s) {
+                          let siphMsg = "siphone l'âme de " + token.get('name') +
+                            " grâce à " + nomPerso(p) + ". " + onGenre(lie, 'Il', 'Elle') + " récupère ";
+                          if (s == soinTotal) {
+                            siphMsg += soin.roll + " pv.";
+                            fraction = 0;
+                          } else {
+                            siphMsg += s + " pv (jet " + soin.roll + ").";
+                            fraction -= Math.ceil(s * 100 / soinTotal);
+                          }
+                          pvMax -= s;
+                          whisperChar(lie.charId, siphMsg);
+                        }, function() {
+                          whisperChar(p.charId, "est déjà au maximum de point de vie. Il laisse échapper l'âme de " + token.get('name'));
+                        });
+                      } else {
+                        whisperChar(p.charId, "est déjà au maximum de point de vie. Il laisse échapper l'âme de " + token.get('name'));
+                      }
                     });
                 });
               }
@@ -3474,7 +3631,7 @@ var COFantasy = COFantasy || function() {
 
   /*
   function logEvents() {
-    var l = eventHistory.length;
+    let l = eventHistory.length;
     log("Historique de taille " + l);
     eventHistory.forEach(function(evt, i) {
       log("evt " + i);
@@ -3546,19 +3703,19 @@ var COFantasy = COFantasy || function() {
   }
 
   function lastEvent() {
-    var l = eventHistory.length;
+    let l = eventHistory.length;
     if (l === 0) return undefined;
     return eventHistory[l - 1];
   }
 
   function setDefaultTokenFromSpec(character, spec, token) {
-    var oldTokenFields = {};
+    let oldTokenFields = {};
     for (const field in spec) {
       if (field.startsWith('_')) continue;
       if (field == 'imgsrc' || field == 'represents' || field == 'top' ||
         field == 'left' || field == 'page_id' || field == 'layer' ||
         field == 'lastmove') continue;
-      var oldValue = token.get(field);
+      let oldValue = token.get(field);
       if (oldValue == spec[field]) continue;
       oldTokenFields[field] = oldValue;
       token.set(field, spec[field]);
@@ -4960,6 +5117,11 @@ var COFantasy = COFantasy || function() {
         if (compagnonPresent(personnage, 'guetteur')) {
           expliquer("Guetteur : +5 en perception");
           bonus += 5;
+        }
+        let lie = personnageAmeLiee(personnage);
+        if (lie && compagnonPresent(lie, 'guetteur')) {
+          expliquer("Guetteur de " + nomPerso(lie) + " : +2 en perception");
+          bonus += 2;
         }
         break;
       case 'persuasion':
@@ -6498,7 +6660,7 @@ var COFantasy = COFantasy || function() {
             error("Il manque un argument à l'option " + args[0], opts);
             return;
           }
-          var plageEC = parseInt(args[1]);
+          let plageEC = parseInt(args[1]);
           if (isNaN(plageEC) || plageEC < 0 || plageEC > 19) {
             error("La plage d'échecs critqiques doit être un nombre positif inférieur à 19", opts);
             return;
@@ -7681,7 +7843,7 @@ var COFantasy = COFantasy || function() {
         case 'drain':
         case 'energie':
           lastType = cmd[0];
-          var l = 0;
+          let l = 0;
           if (scope.additionalDmg) l = scope.additionalDmg.length;
           if (l > 0) {
             scope.additionalDmg[l - 1].type = cmd[0];
@@ -9028,7 +9190,7 @@ var COFantasy = COFantasy || function() {
   function persoInit(perso, evt) {
     let initDerivee = charAttribute(perso.charId, 'initiativeDeriveeDe');
     if (initDerivee.length > 0) {
-      var charDerive = findObjs({
+      let charDerive = findObjs({
         _type: 'character',
         name: initDerivee[0].get('current')
       });
@@ -9093,7 +9255,7 @@ var COFantasy = COFantasy || function() {
     // Sixième sens en sort
     if (attributeAsBool(perso, 'sixiemeSens')) init += 2;
     // Voie du chef d'armée rangs 2 et 3 (Capitaine)
-    var bonusCapitaine = aUnCapitaine(perso, evt);
+    let bonusCapitaine = aUnCapitaine(perso, evt);
     if (bonusCapitaine) init += parseInt(bonusCapitaine);
     if (predicateAsBool(perso, 'graceFeline')) {
       init += modCarac(perso, 'charisme');
@@ -9103,6 +9265,9 @@ var COFantasy = COFantasy || function() {
     }
     if (attributeAsBool(perso, 'masqueDuPredateur')) {
       init += getValeurOfEffet(perso, 'masqueDuPredateur', modCarac(perso, 'sagesse'));
+    }
+    if (attributeAsBool(perso, 'masqueDuPredateurAmeLiee')) {
+      init += getValeurOfEffet(perso, 'masqueDuPredateurAmeLiee', 1);
     }
     if (predicateAsBool(perso, 'controleDuMetabolisme')) {
       init += getValeurOfEffet(perso, 'controleDuMetabolisme', modCarac(perso, 'charisme'));
@@ -9526,6 +9691,12 @@ var COFantasy = COFantasy || function() {
       if (masqueIntense)
         removeTokenAttr(personnage, 'masqueDuPredateurTempeteDeManaIntense', evt);
     }
+    if (attributeAsBool(personnage, 'masqueDuPredateurAmeLiee')) {
+      let bonusMasque =
+        getValeurOfEffet(personnage, 'masqueDuPredateurAmeLiee', 1);
+      attBonus += bonusMasque;
+      explications.push("Masque du prédateur lié : +" + bonusMasque + " en Attaque et DM");
+    }
     if (attributeAsBool(personnage, 'armeSecreteBarde')) {
       attBonus -= 10;
       explications.push("Déstabilisé par une action de charme => -10 en Attaque");
@@ -9571,37 +9742,6 @@ var COFantasy = COFantasy || function() {
     let fio = tm.fioleDeLumiere;
     if (fio === undefined || fio.porteur === undefined) return false;
     return distanceCombat(fio.porteur.token, perso.token) < fio.distance;
-  }
-
-  function compagnonPresent(personnage, attribut) {
-    let attrs = findObjs({
-      _type: 'attribute',
-      _characterid: personnage.charId,
-      name: attribut
-    });
-    if (attrs.length > 0) {
-      let compagnon = attrs[0].get('current');
-      let compToken = findObjs({
-        _type: 'graphic',
-        _subtype: 'token',
-        _pageid: personnage.token.get('pageid'),
-        layer: 'objects',
-        name: compagnon
-      });
-      let res = false;
-      compToken.forEach(function(tok) {
-        if (res) return;
-        let compCharId = tok.get('represents');
-        if (compCharId === '') return;
-        compagnon = {
-          token: tok,
-          charId: compCharId
-        };
-        res = isActive(compagnon);
-      });
-      return res;
-    }
-    return false;
   }
 
   //evt est optionnel
@@ -9704,8 +9844,8 @@ var COFantasy = COFantasy || function() {
       defense += getValeurOfEffet(target, 'aspectDuDemon', 2);
     }
     if (attributeAsBool(target, 'peauDEcorce')) {
-      var bonusPeau = getValeurOfEffet(target, 'peauDEcorce', 1, 'voieDesVegetaux');
-      var peauIntense = attributeAsInt(target, 'peauDEcorceTempeteDeManaIntense', 0);
+      let bonusPeau = getValeurOfEffet(target, 'peauDEcorce', 1, 'voieDesVegetaux');
+      let peauIntense = attributeAsInt(target, 'peauDEcorceTempeteDeManaIntense', 0);
       bonusPeau += peauIntense;
       if (reglesOptionelles.divers.val.forme_d_arbre_amelioree.val && formeDarbre) {
         bonusPeau = Math.ceil(bonusPeau * 1.5);
@@ -9714,6 +9854,11 @@ var COFantasy = COFantasy || function() {
       explications.push("Peau d'écorce : +" + bonusPeau + " en DEF");
       if (peauIntense && evt && !options.test)
         removeTokenAttr(target, 'peauDEcorceTempeteDeManaIntense', evt);
+    }
+    if (attributeAsBool(target, 'peauDEcorceAmeLiee')) {
+      let bonus = getValeurOfEffet(target, 'peauDEcorceAmeLiee', 1);
+      defense += bonus;
+      explications.push("Peau d'écorce liée : +" + bonus + " en DEF");
     }
     if (attributeAsBool(target, 'champDeProtection')) {
       var bonusChamp =
@@ -12827,6 +12972,12 @@ var COFantasy = COFantasy || function() {
         if (predicateAsBool(target, 'armureProtection') && ficheAttributeAsBool(target, 'defarmureon', false)) {
           expliquer("L'armure de protection de " + nomPerso(target) + " le protège du critique");
           diviseDmg++;
+        } else if (predicateAsBool(target, 'bouclierProtection') && ficheAttributeAsInt(target, 'defbouclieron', 0)) {
+          expliquer("Le bouclier de protection de " + nomPerso(target) + " le protège du critique");
+          diviseDmg++;
+        } else if (predicateAsBool(target, 'anneauProtection')) {
+          expliquer("L'anneau de protection de " + nomPerso(target) + " le protège du critique");
+          diviseDmg++;
         }
         if (predicateAsBool(target, 'bouclierProtection') && ficheAttributeAsBool(target, 'defbouclieron', false)) {
           expliquer("Le bouclier de protection de " + nomPerso(target) + " le protège du critique");
@@ -13379,7 +13530,7 @@ var COFantasy = COFantasy || function() {
       if (estMook) {
         error("Les munitions ne sont pas supportées pour les tokens qui ne sont pas liées à un personnage", attackingToken);
       }
-      var munitionsAttr = findObjs({
+      let munitionsAttr = findObjs({
         _type: 'attribute',
         _characterid: attackingCharId,
         name: 'munition_' + options.munition.nom
@@ -13389,7 +13540,7 @@ var COFantasy = COFantasy || function() {
         return;
       }
       munitionsAttr = munitionsAttr[0];
-      var munitions = munitionsAttr.get('current');
+      let munitions = munitionsAttr.get('current');
       if (munitions < 1 || (options.tirDouble && munitions < 2)) {
         sendPerso(attaquant,
           "ne peut pas utiliser cette attaque, car " + sujetAttaquant +
@@ -13658,16 +13809,16 @@ var COFantasy = COFantasy || function() {
               explications.push(
                 weaponName + " explose ! L'arme est complètement détruite");
               sendChat("", "[[2d6]]", function(res) {
-                var rolls = res[0];
-                var explRoll = rolls.inlinerolls[0];
-                var r = {
+                let rolls = res[0];
+                let explRoll = rolls.inlinerolls[0];
+                let r = {
                   total: explRoll.results.total,
                   type: 'normal',
                   display: buildinline(explRoll, 'normal')
                 };
                 dealDamage(attaquant, r, [], evt, false, options, explications,
-                  function(dmgDisplay, dmg) {
-                    var dmgMsg = "<b>Dommages pour " + attackerTokName + " :</b> " +
+                  function(dmgDisplay, dmg, dmDrains) {
+                    let dmgMsg = "<b>Dommages pour " + attackerTokName + " :</b> " +
                       dmgDisplay;
                     addLineToFramedDisplay(display, dmgMsg);
                     finaliseDisplay(display, explications, evt, attaquant, cibles, options);
@@ -13686,7 +13837,7 @@ var COFantasy = COFantasy || function() {
                   display: buildinline(explRoll, 'normal')
                 };
                 dealDamage(attaquant, r, [], evt, false, options, explications,
-                  function(dmgDisplay, dmg) {
+                  function(dmgDisplay, dmg, dmDrains) {
                     var dmgMsg =
                       "<b>Dommages pour " + attackerTokName + " :</b> " +
                       dmgDisplay;
@@ -14988,6 +15139,11 @@ var COFantasy = COFantasy || function() {
       let bonusMasque = getValeurOfEffet(attaquant, 'masqueDuPredateur', modCarac(attaquant, 'sagesse'));
       if (bonusMasque > 0) attDMBonusCommun += " +" + bonusMasque;
     }
+    if (attributeAsBool(attaquant, 'masqueDuPredateurAmeLiee')) {
+      let bonusMasque =
+        getValeurOfEffet(attaquant, 'masqueDuPredateurAmeLiee', 1);
+      if (bonusMasque > 0) attDMBonusCommun += " +" + bonusMasque;
+    }
     if (options.bonusDM) {
       attDMBonusCommun += " +" + options.bonusDM;
     }
@@ -15396,12 +15552,15 @@ var COFantasy = COFantasy || function() {
               } else {
                 target.messages.push("Attaque sournoise => +" + sournoise + options.d6 + " DM");
               }
-              var valueSournoise = sournoise + options.d6;
+              let valueSournoise = sournoise + options.d6;
               if (predicateAsBool(target, 'armureProtection') && ficheAttributeAsBool(target, 'defarmureon', false)) {
                 target.messages.push("L'armure de protection de " + nomPerso(target) + " réduit l'attaque sournoise");
                 valueSournoise = "ceil(" + valueSournoise + "/2)";
               } else if (predicateAsBool(target, 'bouclierProtection') && ficheAttributeAsInt(target, 'defbouclieron', 0)) {
                 target.messages.push("Le bouclier de protection de " + nomPerso(target) + " réduit l'attaque sournoise");
+                valueSournoise = "ceil(" + valueSournoise + "/2)";
+              } else if (predicateAsBool(target, 'anneauProtection')) {
+                target.messages.push("L'anneau de protection de " + nomPerso(target) + " réduit l'attaque sournoise");
                 valueSournoise = "ceil(" + valueSournoise + "/2)";
               }
               target.additionalDmg.push({
@@ -15968,9 +16127,22 @@ var COFantasy = COFantasy || function() {
                       let pcVampirise = target.vampirise || options.vampirise;
                       let soinsVamp = dmgDrain || 0;
                       if (pcVampirise) soinsVamp += Math.ceil(dmg * pcVampirise / 100);
+                      let lie = personnageAmeLiee(attaquant);
                       soigneToken(attaquant, soinsVamp, evt, function(soins) {
                         target.messages.push(
                           "L'attaque soigne " + attackerTokName + " de " + soins + " PV");
+                        if (lie && (soins > 4 || soins < soinsVamp)) {
+                          let soin2 = soinsVamp - soins + Math.floor(soins / 5);
+                          soigneToken(lie, soin2, evt, function(s) {
+                            target.messages.push("L'attaque soigne aussi " + nomPerso(lie) + " de " + s + " PV");
+                          });
+                        }
+                      }, function() {
+                        if (lie) {
+                          soigneToken(lie, soinsVamp, evt, function(soins) {
+                            target.messages.push("L'attaque soigne " + nomPerso(lie) + " de " + soins + " PV");
+                          });
+                        }
                       });
                     }
                     let absorptionEnergie = predicateAsInt(attaquant, 'absorptionEnergie', 0);
@@ -17360,12 +17532,12 @@ var COFantasy = COFantasy || function() {
           }
         }, expliquer, options);
     }
-    var dmSuivis = {
+    let dmSuivis = {
       drain: 0
     }; //si il faut noter les DMs d'un type particulier
     if (mainDmgType == 'drain') dmSuivis.drain = dmgTotal;
     charAttribute(target.charId, 'vitaliteSurnaturelle').forEach(function(a) {
-      var typeVitalite = a.get('max').split(',');
+      let typeVitalite = a.get('max').split(',');
       typeVitalite.forEach(function(tv) {
         if (tv == mainDmgType) dmSuivis[tv] = dmgTotal;
         else dmSuivis[tv] = 0;
@@ -17374,7 +17546,7 @@ var COFantasy = COFantasy || function() {
     // Autres sources de dégâts
     // On compte d'abord les autres sources, pour la synchronisation
     let count = 0;
-    for (var dt in dmgParType) {
+    for (let dt in dmgParType) {
       if (immuniseAuType(target, dt, options.attaquant)) {
         if (expliquer && !target['msgImmunite_' + dt]) {
           expliquer(nomPerso(target) + " ne semble pas affecté par " + stringOfType(dt));
@@ -17844,6 +18016,7 @@ var COFantasy = COFantasy || function() {
           }
         }
         if (attributeAsBool(target, 'masqueMortuaire')) rd += 2;
+        if (attributeAsBool(target, 'masqueMortuaireAmeLiee')) rd += 1;
         if (rdTarget.nature > 0 && dmgNaturel(options)) rd += rdTarget.nature;
         if (dmgTotal > rd && rdTarget.sauf[1]) {
           if (dmgTotal > rd + rdTarget.sauf[1]) rd += rdTarget.sauf[1];
@@ -26313,7 +26486,9 @@ var COFantasy = COFantasy || function() {
 
   function estNonVivant(perso) {
     return (predicateAsBool(perso, 'nonVivant') ||
-      attributeAsBool(perso, 'masqueMortuaire') || estMortVivant(perso));
+      attributeAsBool(perso, 'masqueMortuaire') ||
+      attributeAsBool(perso, 'masqueMortuaireAmeLiee') ||
+      estMortVivant(perso));
   }
 
   function estElfeNoir(perso) {
@@ -38845,6 +39020,12 @@ var COFantasy = COFantasy || function() {
       fin: "retrouve une peau normale",
       visible: true
     },
+    peauDEcorceAmeLiee: {
+      activation: "voit sa peau se durcir",
+      actif: "a la peau un peu rugueuse",
+      fin: "retrouve une peau normale",
+      visible: true
+    },
     rayonAffaiblissant: {
       activation: "est touché par un rayon affaiblissant",
       activationF: "est touchée par un rayon affaiblissant",
@@ -39277,6 +39458,12 @@ var COFantasy = COFantasy || function() {
       fin: "retrouve une apparence de vivant",
       visible: true
     },
+    masqueMortuaireAmeLiee: {
+      activation: "est lié à une apparence de la mort",
+      activationF: "est liée à une apparence de la mort",
+      actif: "semble en lien avec la mort",
+      fin: "n'est plus en lien avec la mort"
+    },
     armeBrulante: {
       activation: "sent son arme lui chauffer la main",
       actif: "se brûle la main sur son arme",
@@ -39294,6 +39481,13 @@ var COFantasy = COFantasy || function() {
       actif: "a les traits d'un prédateur",
       fin: "redevient normal",
       visible: true
+    },
+    masqueDuPredateurAmeLiee: {
+      activation: "est lié à une âme de prédateur",
+      activationF: "est liée à une âme de prédateur",
+      actif: "bénéficie d'un lien avec un prédateur",
+      fin: "le lien disparaît",
+      visible: false
     },
     aspectDeLaSuccube: {
       activation: "acquiert une beauté fascinante",
@@ -43145,7 +43339,7 @@ on('ready', function() {
     //Ensuite les prédicats booléens
     let predicates = '^(' +
       'actionLibre|agripper|ambidextreDuelliste|animal|argumentDeTaille' +
-      '|armureProtection|armureProtection|aucuneActionCombat|baroudHonneur' +
+      '|armureProtection|aucuneActionCombat|baroudHonneur' +
       '|botteMortelle|bouclierPsi|briseurDOs|bucheron|champion' +
       '|chasseurEmerite|chatimentDuMale|chimiste|ciblesMultiples' +
       '|combatEnPhalange|combatKinetique|commandant|controleDuMetabolisme' +